+++ src/XUnity.AutoTranslator.Plugin.BepIn/AutoTranslatorPlugin.cs
@@ -39,6 +39,8 @@ namespace XUnity.AutoTranslator.Plugin.BepIn
 
       public string ConfigPath => _dataPath;
 
+      public bool AllowRuntimeHooksByDefault => false;
+
       public IniFile ReloadConfig()
       {
          if( !File.Exists( _configPath ) )