+++ src/XUnity.AutoTranslator.Plugin.Core/DefaultPluginEnvironment.cs
@@ -33,6 +33,8 @@ namespace XUnity.AutoTranslator.Plugin.Core
 
       public string ConfigPath => _dataFolder;
 
+      public bool AllowRuntimeHooksByDefault => true;
+
       public void SaveConfig()
       {
          _file.Save( _configPath );